+++ src/main/java/org/c4marathon/assignment/service/AccountSchedulerService.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.c4marathon.assignment.repository.MainAccountRepository;
+import org.c4marathon.assignment.repository.mainAccount.MainAccountRepository;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
